make ids Long

+++ src/main/scala/com/dblazejewski/domain/IdEntity.scala
@@ -1,5 +1,5 @@
 package com.dblazejewski.domain
 
 trait IdEntity {
-  def id: Option[Int]
+  def id: Option[Long]
 }
