+++ src/main/scala/com/dblazejewski/domain/Post.scala
@@ -2,8 +2,8 @@ package com.dblazejewski.domain
 
 import java.time.LocalDateTime
 
-case class Post(id: Option[Int],
-                authorId: Int,
-                groupId: Int,
+case class Post(id: Option[Long],
+                authorId: Long,
+                groupId: Long,
                 createdAt: LocalDateTime,
                 content: String) extends IdEntity
