+++ src/main/scala/com/dblazejewski/repository/UserRepository.scala
@@ -17,7 +17,7 @@ trait UserSchema {
   import database.driver.api._
 
   class UserTable(tag: slick.lifted.Tag) extends Table[User](tag, "USER") {
-    def id: Rep[Int] = column[Int]("id", O.PrimaryKey, O.AutoInc)
+    def id: Rep[Long] = column[Long]("id", O.PrimaryKey, O.AutoInc)
 
     def name: Rep[String] = column[String]("name")
